backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
python-dotenv
google-generativeai
pydantic
requests
